requirements.txt
Django==2.1.7
djangorestframework==3.9.1
djangorestframework-filters==1.0.0.dev0
django-cors-headers==2.4.0
django-rest-swagger==2.1.2
Pillow==5.4.1
mysqlclient==1.4.2